+++ src/bin/disco/download.rs
@@ -14,11 +14,12 @@ pub enum Dataset {
     Movielens1m,
     Movielens25m,
     MovielensLatestSmall,
+    MovielensLatest,
 }
 
 impl Dataset {
-    pub fn variants() -> [&'static str; 4] {
-        ["movielens-100k", "movielens-1m", "movielens-25m", "movielens-latest-small"]
+    pub fn variants() -> [&'static str; 5] {
+        ["movielens-100k", "movielens-1m", "movielens-25m", "movielens-latest-small", "movielens-latest"]
     }
 }
 
@@ -31,6 +32,7 @@ impl FromStr for Dataset {
             "movielens-1m" => Ok(Dataset::Movielens1m),
             "movielens-25m" => Ok(Dataset::Movielens25m),
             "movielens-latest-small" => Ok(Dataset::MovielensLatestSmall),
+            "movielens-latest" => Ok(Dataset::MovielensLatest),
             // not shown since possible_values used
             _ => Err(format!("Invalid dataset: {}", s)),
         }
@@ -44,6 +46,7 @@ impl ToString for Dataset {
             Dataset::Movielens1m => "movielens-1m".to_string(),
             Dataset::Movielens25m => "movielens-25m".to_string(),
             Dataset::MovielensLatestSmall => "movielens-latest-small".to_string(),
+            Dataset::MovielensLatest => "movielens-latest".to_string(),
         }
     }
 }
@@ -274,6 +277,55 @@ fn download_movielens_latest_small(output: &Path, overwrite: bool) -> Result<(),
     Ok(())
 }
 
+fn download_movielens_latest(output: &Path, overwrite: bool) -> Result<(), Box<dyn Error>> {
+    let mut movies = HashMap::new();
+
+    let archive_data = download_file(
+        "https://files.grouplens.org/datasets/movielens/ml-latest.zip",
+        "b9c23b665ee348bd1fadfadca688b8750c575f3af76d3441cd50cba87ad2c4df",
+    )?;
+    let cursor = Cursor::new(archive_data);
+    let mut archive = zip::ZipArchive::new(cursor)?;
+
+    // make borrow checker happy
+    {
+        let movies_data = archive.by_name("ml-latest/movies.csv")?;
+        let mut rdr = csv::ReaderBuilder::new()
+            .from_reader(movies_data);
+        for result in rdr.records() {
+            let record = result?;
+            let id = record.get(0).unwrap().to_string();
+            let title = record.get(1).unwrap().to_string();
+            movies.insert(id, title);
+        }
+    }
+
+    let file = create_file(output, overwrite)?;
+    let mut wtr = csv::Writer::from_writer(file);
+    wtr.write_record(&["user_id", "item_id", "rating"])?;
+
+    // show processing progress since it takes a while
+    let bar = progress_bar(27753444, "Processing", "{msg} {wide_bar} {percent}%");
+    bar.set_draw_delta(250000);
+
+    let ratings_data = archive.by_name("ml-latest/ratings.csv")?;
+    let mut rdr = csv::ReaderBuilder::new()
+        .from_reader(ratings_data);
+    for result in rdr.records() {
+        let record = result?;
+        let user_id = record.get(0).unwrap().to_string();
+        let item_id = record.get(1).unwrap().to_string();
+        let rating = record.get(2).unwrap().to_string();
+        wtr.write_record(&[user_id, movies.get(&item_id).unwrap().to_string(), rating])?;
+        bar.inc(1);
+    }
+
+    wtr.flush()?;
+    bar.finish();
+
+    Ok(())
+}
+
 pub fn download(dataset: Dataset, output: Option<PathBuf>, overwrite: bool) -> Result<(), Box<dyn Error>> {
     let output = output.unwrap_or_else(|| {
         let mut default_output = PathBuf::from(&dataset.to_string());
@@ -289,6 +341,7 @@ pub fn download(dataset: Dataset, output: Option<PathBuf>, overwrite: bool) -> R
         Dataset::Movielens1m => "https://files.grouplens.org/datasets/movielens/ml-1m-README.txt",
         Dataset::Movielens25m => "https://files.grouplens.org/datasets/movielens/ml-25m-README.html",
         Dataset::MovielensLatestSmall => "https://files.grouplens.org/datasets/movielens/ml-latest-small-README.html",
+        Dataset::MovielensLatest => "https://files.grouplens.org/datasets/movielens/ml-latest-README.html",
     };
     eprintln!("For dataset usage info, see {}", usage_url);
 
@@ -297,6 +350,7 @@ pub fn download(dataset: Dataset, output: Option<PathBuf>, overwrite: bool) -> R
         Dataset::Movielens1m => download_movielens_1m(&output, overwrite),
         Dataset::Movielens25m => download_movielens_25m(&output, overwrite),
         Dataset::MovielensLatestSmall => download_movielens_latest_small(&output, overwrite),
+        Dataset::MovielensLatest => download_movielens_latest(&output, overwrite),
     };
     if res.is_ok() {
         eprintln!("Saved to {}", output.display());