+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 0.1.1 (unreleased)
 
+- Added `movielens-latest` dataset
 - Added `movielens-latest-small` dataset
 
 ## 0.1.0 (2021-12-01)